Chore: 지도 기능 디렉터리 구조 개편
