+++ src/features/map/api/mapApi.ts
@@ -1,3 +1,4 @@
+import { apiClient } from "@/shared/api/client";
 import type {
   CategoryResponse,
   DistrictResponse,
@@ -8,16 +9,14 @@ import type {
   StoreResponse,
 } from "@/types/mapTypes";
 
-import { apiClient } from "./client";
-
 /**
- * LocationParams 객체를 URL 쿼리스트링 형태로 변환합니다.
+ * LocationParams 객체를 URL 쿼리스트링으로 변환합니다.
  *
- * 예: { province: "SEOUL", district: "GANGNAM" }
- *  → "?province=SEOUL&district=GANGNAM"
- *
- * @param params - 도/시군구 정보를 담은 객체 (없으면 빈 문자열 반환)
- * @returns 쿼리스트링이 포함된 문자열 (예: "?province=SEOUL&district=GANGNAM")
+ * 예:
+ * ```ts
+ * buildQuery({ province: "SEOUL", district: "GANGNAM" })
+ * // => "?province=SEOUL&district=GANGNAM"
+ * ```
  */
 const buildQuery = (params?: LocationParams) => {
   if (!params) return "";
@@ -43,26 +42,30 @@ const buildQuery = (params?: LocationParams) => {
 export const getProvinces = () => apiClient<ProvinceResponse>("/api/map/provinces");
 
 /**
- * province 값을 전달하면 해당 도의 시/구/군 목록만 반환합니다.
+ * 특정 도에 속한 시/군/구 목록을 조회합니다.
+ *
+ * @param params.province - 도/광역시 ID
  */
 export const getDistricts = (params?: Pick<LocationParams, "province">) =>
   apiClient<DistrictResponse>(`/api/map/districts${buildQuery(params)}`);
 
 /**
- * 시/군/구와 도 정보를 전달하면 해당 읍/면/동 목록을 반환합니다.
+ * 시/군/구와 도 정보를 기반으로 읍/면/동 목록을 조회합니다.
+ *
+ * @param params - province와 district를 포함한 객체
  */
 export const getNeighborhoods = (params?: LocationParams) =>
   apiClient<NeighborhoodResponse>(`/api/map/neighborhoods${buildQuery(params)}`);
 
 /**
- * 지도에서 사용할 카테고리 목록을 조회합니다.
+ * 지도에서 사용 가능한 카테고리 목록을 조회합니다.
  */
 export const getCategories = () => apiClient<CategoryResponse>("/api/map/categories");
 
 /**
  * 가게 목록을 조회합니다.
  *
- * @param filters - 지역/카테고리/좌표 정보를 담은 객체
+ * @param filters - 지역, 카테고리, 좌표 정보를 담은 객체
  */
 export const getStores = (filters?: StoreFilters) => {
   const params = new URLSearchParams();
