+++ src/features/map/search/ui/MapButtonGroup.tsx
@@ -3,11 +3,18 @@
 import { useDispatch, useSelector } from "react-redux";
 
 import Button from "@/components/common/button/Button";
-import { useCategories } from "@/shared/hooks/useCategories";
 import type { RootState } from "@/shared/store";
 import { setSelectedCategory } from "@/shared/store/mapSlice";
 import type { FilterCategory } from "@/types/mapTypes";
 
+import { useCategories } from "../api/useCategories";
+
+/**
+ * 지도에서 사용할 카테고리 버튼을 렌더링합니다.
+ *
+ * - React Query로 카테고리 목록을 가져온 뒤 버튼으로 노출합니다.
+ * - 선택 결과는 Redux `selectedCategory`에 저장되어 필터링에 활용됩니다.
+ */
 export default function MapButtonGroup() {
   const dispatch = useDispatch();
   const activeCategoryCode = useSelector((state: RootState) => state.map.selectedCategory);
@@ -15,7 +22,6 @@ export default function MapButtonGroup() {
 
   const categories: FilterCategory[] = data?.data ?? [];
 
-  // 버튼을 누르면 선택된 카테고리를 전역 상태에 저장합니다.
   const handleCategoryClick = (code: string) => {
     dispatch(setSelectedCategory(activeCategoryCode === code ? null : code));
   };
